Added OUTBOUND_UDP_TIMEOUT to config file.
Adding the OUTBOUND_UDP_TIMEOUT to the config file.  This is the
timeout for the UDP socket that is used to send outbound messages.
This will help with potential locking that UDP sockets can cause.

In testing I did see a rare side case that resulted from all the
workers locking when their was a network blip. This should help
with that and recover faster.